backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
python-dotenv==0.19.0
pip freeze > requirements.txt